Chore : batch read 속도 테스트 클래스, 패키지 분리
